Database Operation for sale's line item.

+++ src/com/refresh/pos/core/LineItem.java
@@ -14,7 +14,7 @@ public class LineItem {
 		this.quantity = quantity;
 	}
 
-	public long getProductId() {
+	public int getProductId() {
 		return product.getId();
 	}
